% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aws-download.R
\name{aws_s3_download}
\alias{aws_s3_download}
\title{download files or folders to AWS}
\usage{
aws_s3_download(path, bucket, key, check = TRUE, error = FALSE)
}
\arguments{
\item{path}{String. Path to download location. Could be a folder or specific file(s)}

\item{bucket}{The name of the bucket to be downloaded from}

\item{key}{The key or name for the file or folder to be download from the bucket.
Should end with "/" for folders. Use "" to download whole bucket.}

\item{check}{Whether to check if the exact file already exists in the download location
and skip downloading. Defaults to TRUE}

\item{error}{Whether error out if the file is missing, folder is empty, or
system environment variables are missing. Otherwise a message will print
but an empty list will be returned.}
}
\value{
A list, each element  having the key, etag (hash), and path of downloaded
files
}
\description{
Providing a key that is a folder and path that is a folder will result in
the whole folder being copied to the path location. If you supply "" to key,
the whole bucket will be downloaded.
}
