% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gather-file-paths.R
\name{create_git_prefix}
\alias{create_git_prefix}
\title{Create git prefix for file paths}
\usage{
create_git_prefix(
  branch = gert::git_branch(),
  path = "",
  ignore_if_main = TRUE
)
}
\arguments{
\item{branch}{String. Current branch of repo.}

\item{path}{String. Folder where file should be uploaded. If an empty string
is provided (default), the path argument will be ignored.}

\item{ignore_if_main}{Logical. No prefixed added to files on main branch.}
}
\value{
String. Prefix for the file path.
}
\description{
Adds information about the git branch to the file path.
}
\examples{

# create_git_prefix(branch = gert::git_branch(),
# path= "example",
# ignore_if_main=TRUE)
# "refs/head/current/branch/example"

}
