% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gather-file-paths.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Get html report files}
\usage{
get_file_paths(tar_obj, pattern, ...)
}
\arguments{
\item{tar_obj}{a tar_render object or other target object that contains file paths}

\item{pattern}{string. regex pattern to match desired file paths}

\item{...}{additional tar_render or other target objects}
}
\value{
a list of file paths
}
\description{
Get html report files
}
