% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-make-urls.R
\name{make_eha_secure_url}
\alias{make_eha_secure_url}
\title{Make urls for EHA secure}
\usage{
make_eha_secure_url(url_domain = Sys.getenv("URL_PREFIX"), remote_path)
}
\arguments{
\item{url_domain}{String. Resource name on eha secure}

\item{remote_path}{String. What is the path to the resource on remote site. For
AWS this will be the item key.}
}
\value{
string. URL for resource
}
\description{
Create urls for items stored on EHA secure sites
}
