% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-send-email.R
\name{send_email_update}
\alias{send_email_update}
\title{Send email alerts to updated automation reports}
\usage{
send_email_update(
  to,
  from = "rsconnect@ecohealthalliance.org",
  project_name,
  attach = FALSE,
  test = FALSE
)
}
\arguments{
\item{to}{A vector of email addresses serving as primary recipients for the
message.}

\item{from}{The email address of the sender.}

\item{project_name}{String. Name of the project to use in email subject and body text.}

\item{attach}{Logical. Should reports be attached to email? Default is
FALSE. If TRUE, all HTML reports found in \emph{"outputs"} folder will be
attached to email.}

\item{test}{Logical. Is this an email alert for testing automation reports?
Default is FALSE.}
}
\value{
Invisible. Update email sent to list of recipients in \code{to}
}
\description{
Send email alerts to updated automation reports
}
