% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aws-upload.R
\name{aws_s3_upload}
\alias{aws_s3_upload}
\title{Upload files or folders to AWS}
\usage{
aws_s3_upload(
  path,
  bucket,
  key = basename(path),
  prefix = "",
  check = TRUE,
  error = FALSE
)
}
\arguments{
\item{path}{The path to the file(s) or folder(s) to be uploaded}

\item{bucket}{The name of the bucket to be uploaded to}

\item{key}{The key or name for the file or folder to take in the bucket.
Should end with "/" for folders. Use "" to upload files in folder without
top-level folder.}

\item{prefix}{A prefix to prepend to the file or folder keys. Generally
should end with "/"}

\item{check}{Whether to check if the exact file already exists in the bucket
and skip uploading. Defaults to TRUE}

\item{error}{Whether error out if the file is missing, folder is empty, or
system environment variables are missing. Otherwise a message will print
but an empty list will be returned.}
}
\value{
A list, each element being having the key and etag (hash) of uploaded
files
}
\description{
Upload files or folders to AWS
}
