% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aws-upload.R
\name{aws_s3_upload}
\alias{aws_s3_upload}
\title{Upload files or folders to AWS}
\usage{
aws_s3_upload(
  path,
  bucket,
  key = basename(path),
  prefix = "",
  check = TRUE,
  error = FALSE,
  file_type = "guess"
)
}
\arguments{
\item{path}{String. The path to the file(s) or folder(s) to be uploaded}

\item{bucket}{String. The name of the bucket to be uploaded to}

\item{key}{String. The key or name for the file or folder to take in the bucket.
Should end with "/" for folders. Use "" (empty string) to upload files in folder without
top-level folder.}

\item{prefix}{String. A prefix to prepend to the file or folder keys. Generally
should end with "/"}

\item{check}{Logical. Whether to check if the exact file already exists in the bucket
and skip uploading. Defaults to TRUE}

\item{error}{Logical. Whether error out if the file is missing, folder is empty, or
system environment variables are missing. Otherwise a message will print
but an empty list will be returned.}

\item{file_type}{String. Provide a file type from \code{\link[mime:mimemap]{mime::mimemap()}} (e.g. "html","csv")
or provide "guess"to call \code{\link[mime:guess_type]{mime::guess_type()}}.}
}
\value{
A list, each element being having the key and etag (hash) of uploaded
files
}
\description{
When uploading folders, the subdirectory structure will be preserved. To
upload files from a folder without preserving the directory structure,
pass a vector of file paths to the path argument.
}
\details{
If you would like the change the directory structure, pass in a vector of
file paths and a corresponding vector of keys.
}
